Renamed TaskStore to useTaskStore.
